pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prufrock"
version = "0.1.0"
description = "Your life. Your voice. Your book. — Memoir pipeline from physical artifacts."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Robert Chuvala", email = "rob@northwoodssentinel.com" }]

dependencies = [
    "click>=8.0",
    "anthropic>=0.40",
    "Pillow>=10.0",
    "face-recognition>=1.3",
    "exifread>=3.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
prufrock = "prufrock.cli:cli"
